cumulative reference changes after http://mercurial.mcs.anl.gov/ad/xaifBooster/rev/a518f20d91fc  and http://mercurial.mcs.anl.gov/ad/OpenADFortTk/rev/f7b2463554ab in particular: - using temporaries helper for accumulation variables and changing the scope of these variables (leads to changed numbering in other temp variables)  done with http://mercurial.mcs.anl.gov/ad/xaifBooster/rev/6e882612cc26 - renaming the conversion routines to oad_convert done with http://mercurial.mcs.anl.gov/ad/xaifBooster/rev/017697be821a

+++ testRoundTrip/examples/arrayBoundVar/refOutput/admsplithead_sf.pre.xb.x2w.w2f.f
@@ -61,11 +61,11 @@ C     $OpenAD$ END REPLACEMENT
 C     $OpenAD$ BEGIN REPLACEMENT 2
 C$OPENAD XXX Template ad_template.f
       __value__(Z(1 : INT(SIZE(X)))) = __value__(X(1 : 2))
-      OpenAD_Symbol_0 = SIZE(INT(__value__(X)))
+      OpenAD_Symbol_0 = SIZE(__value__(X))
 C     $OpenAD$ INLINE push_i(subst)
       CALL push_i(OpenAD_Symbol_0)
       __value__(Y(1 : 2)) = __value__(Z(1 : INT(SIZE(X))))
-      OpenAD_Symbol_1 = SIZE(INT(__value__(X)))
+      OpenAD_Symbol_1 = SIZE(__value__(X))
 C     $OpenAD$ INLINE push_i(subst)
       CALL push_i(OpenAD_Symbol_1)
 C     $OpenAD$ END REPLACEMENT
@@ -100,11 +100,11 @@ C     $OpenAD$ END REPLACEMENT
 C     $OpenAD$ BEGIN REPLACEMENT 10
 C$OPENAD XXX Template ad_template.f
       __value__(Z(1 : INT(SIZE(X)))) = __value__(X(1 : 2))
-      OpenAD_Symbol_0 = SIZE(INT(__value__(X)))
+      OpenAD_Symbol_0 = SIZE(__value__(X))
 C     $OpenAD$ INLINE push_i(subst)
       CALL push_i(OpenAD_Symbol_0)
       __value__(Y(1 : 2)) = __value__(Z(1 : INT(SIZE(X))))
-      OpenAD_Symbol_1 = SIZE(INT(__value__(X)))
+      OpenAD_Symbol_1 = SIZE(__value__(X))
 C     $OpenAD$ INLINE push_i(subst)
       CALL push_i(OpenAD_Symbol_1)
 C     $OpenAD$ END REPLACEMENT
